Create README.md
using this YouTube videos with appropriate URLs can be downloaded. this has also a facility to convert the video into audio and then download it. this will ask the user about the directory in which the user wants to save the downloaded files. it gives the facility to choose the stream according to which the user want to download the audio or the video.

this may not work properly due to the recent changes in pytube package.